Extract jexunit-base module (Phase 1 refactoring)
Move all JUnit-agnostic code from core into a new jexunit-base module:
- ExcelLoader, model, data (OGNL), context, SPI, config, command dispatch
- Introduce TestCommandHost interface to decouple TestCommandRunner from JExUnitBase
- Replace @RunWith(JExUnit.class) check in TestCommandScanner with @ExcelFile field/method scan
- Replace Assert.fail() in CommandValidator with throw new AssertionError
- core now depends on jexunit-base and retains only the 3 JUnit 4 runner classes
