--- core/src/main/java/com/jexunit/core/commands/package-info.java
@@ -1,8 +1,0 @@
-/**
- * This package is used for defining and finding the test-commands. Test-command preparation and execution will also be
- * handled here.
- * 
- * @author fabian
- *
- */
-package com.jexunit.core.commands;
