--- core/src/main/java/com/jexunit/core/package-info.java
@@ -1,21 +1,0 @@
-/**
- * In this package you will find the main-part for integrating JExUnit.
- * <p>
- * Running your tests using JExUnit:<br>
- * <br>
- * <code>
- * &#064;RunWith(JexUnit.class)<br>
- * public class MyTest {<br>
- *     &#064;ExcelFile<br>
- *     private static String myExcelFile = "path-to-my-excel-file.xlsx";<br>
- * }
- * </code>
- * </p>
- * <p>
- * To change the default configuration, just add a <i>jexunit.properties</i> file to the classpath.
- * </p>
- * 
- * @author fabian
- *
- */
-package com.jexunit.core;
