+++ jexunit-base/src/main/java/com/jexunit/core/JExUnitConfig.java
