+++ jexunit-base/src/main/java/com/jexunit/core/commands/Command.java
