+++ jexunit-base/src/main/java/com/jexunit/core/commands/TestCommandHost.java
@@ -1,0 +1,8 @@
+package com.jexunit.core.commands;
+
+import com.jexunit.core.model.TestCase;
+
+public interface TestCommandHost {
+    Class<?> getTestType();
+    void runCommand(TestCase<?> testCase) throws Exception;
+}
