+++ jexunit-base/src/main/java/com/jexunit/core/data/CollectionPropertyHelper.java
