+++ jexunit-base/src/main/java/com/jexunit/core/data/package-info.java
@@ -1,7 +1,7 @@
 /**
  * Package for handling the data (read the excel file(s),
  * create/modify the entity-objetcs, ...).
- * 
+ *
  * @author fabian
  *
  */
