+++ jexunit-base/src/main/java/com/jexunit/core/model/package-info.java
@@ -1,6 +1,6 @@
 /**
  * This package contains the model classes representing the internal data structure.
- * 
+ *
  * @author fabian
  *
  */
